Added bunch of logic and classes to calculate cost of an order

+++ src/main/java/app/persistence/OrderDetails.java
@@ -2,7 +2,7 @@ package app.persistence;
 
 import app.entities.OrderStatus;
 
-public record OrderDetails(String remark, OrderStatus status, Integer price) {
+public record OrderDetails(String remark, OrderStatus status, Double price) {
 
     public OrderDetails(String remark, OrderStatus status) {
         this(remark, status, null);
